Home screen and removal of old screens
