+++ project.properties
@@ -8,4 +8,4 @@
 # project structure.
 
 # Project target.
-target=android-18
+target=android-19
