+++ res/layout/dashboard_brew_group_view_list_child.xml
@@ -7,17 +7,17 @@
         android:id="@+id/brew_group_player_name"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_alignBottom="@+id/imageView1"
+        android:layout_alignBottom="@+id/brew_round_icon"
         android:layout_alignParentLeft="true"
         android:layout_alignParentTop="true"
-        android:layout_toLeftOf="@+id/imageView1"
+        android:layout_toLeftOf="@+id/brew_round_icon"
         android:gravity="center_vertical"
         android:textAlignment="center"
         android:text="Player Name"
         android:textAppearance="?android:attr/textAppearanceMedium" />
 
     <ImageView
-        android:id="@+id/imageView1"
+        android:id="@+id/brew_round_icon"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_alignParentRight="true"
