+++ src/com/morgan/design/activity/refactor/DashbaordActivity.java
@@ -13,23 +13,20 @@ import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentPagerAdapter;
 import android.support.v4.view.ViewPager;
-import android.view.LayoutInflater;
 import android.view.Menu;
-import android.view.View;
-import android.view.View.OnLongClickListener;
-import android.view.ViewGroup;
-import android.widget.AdapterView;
-import android.widget.AdapterView.OnItemLongClickListener;
-import android.widget.ExpandableListAdapter;
-import android.widget.ExpandableListView;
-import android.widget.TextView;
+import android.view.MenuItem;
 
 import com.morgan.design.R;
 import com.morgan.design.activity.BaseBrewFragmentActivity;
-import com.morgan.design.adaptor.refactor.BrewGroupsExpandableListAdapter;
 import com.morgan.design.db.domain.BrewGroup;
 import com.morgan.design.db.domain.BrewStats;
 import com.morgan.design.db.domain.PlayerStats;
+import com.morgan.design.fragment.refactor.BrewGroupsFragment;
+import com.morgan.design.fragment.refactor.BrewHomePageFragments;
+import com.morgan.design.fragment.refactor.BrewStatsFragments;
+import com.morgan.design.helpers.Changelog;
+import com.morgan.design.utils.Prefs;
+import com.morgan.design.utils.Utils;
 
 public class DashbaordActivity extends BaseBrewFragmentActivity implements ActionBar.TabListener {
 
@@ -94,8 +91,28 @@ public class DashbaordActivity extends BaseBrewFragmentActivity implements Actio
 	@Override
 	public boolean onCreateOptionsMenu(Menu menu) {
 		// Inflate the menu; this adds items to the action bar if it is present.
-		getMenuInflater().inflate(R.menu.dashbaord, menu);
+		getMenuInflater().inflate(R.menu.dashboard_menu, menu);
 		return true;
+	}
+
+	@Override
+	public boolean onOptionsItemSelected(final MenuItem item) {
+		switch (item.getItemId()) {
+			case R.id.db_credits:
+				Utils.openCredits(this);
+				return true;
+			case R.id.db_feedback:
+				Utils.openFeedback(this);
+				return true;
+			case R.id.db_settings:
+				Prefs.openPreferenecesActivity(this);
+				return true;
+			case R.id.db_changelog:
+				Changelog.show(this);
+				return true;
+			default:
+				return super.onOptionsItemSelected(item);
+		}
 	}
 
 	@Override
@@ -112,10 +129,6 @@ public class DashbaordActivity extends BaseBrewFragmentActivity implements Actio
 	public void onTabReselected(ActionBar.Tab tab, FragmentTransaction fragmentTransaction) {
 	}
 
-	/**
-	 * A {@link FragmentPagerAdapter} that returns a fragment corresponding to
-	 * one of the sections/tabs/pages.
-	 */
 	public class SectionsPagerAdapter extends FragmentPagerAdapter {
 
 		public SectionsPagerAdapter(FragmentManager fm) {
@@ -124,36 +137,27 @@ public class DashbaordActivity extends BaseBrewFragmentActivity implements Actio
 
 		@Override
 		public Fragment getItem(int position) {
-			// getItem is called to instantiate the fragment for the given page.
-			// Return a DummySectionFragment (defined as a static inner class
-			// below) with the page number as its lone argument.
 			Fragment fragment = null;
-
 			switch (position) {
 				case 0:
 					fragment = new BrewGroupsFragment();
 					((BrewGroupsFragment) fragment).setBrewGroups(groups);
+					break;
+				case 1:
+					fragment = new BrewHomePageFragments();
 					break;
 				case 2:
 					fragment = new BrewStatsFragments();
 					((BrewStatsFragments) fragment).setBrewStats(brewStats);
 					((BrewStatsFragments) fragment).setPlayerStats(playerStats);
 					break;
-				default:
-					fragment = new DummySectionFragment();
-					Bundle args = new Bundle();
-					args.putInt(DummySectionFragment.ARG_SECTION_NUMBER, position + 1);
-					fragment.setArguments(args);
-					break;
 			}
-
 			return fragment;
 		}
 
 		@Override
 		public int getCount() {
-			// Show 3 total pages.
-			return 3;
+			return 3; // Show 3 total pages.
 		}
 
 		@Override
@@ -168,113 +172,6 @@ public class DashbaordActivity extends BaseBrewFragmentActivity implements Actio
 					return getString(R.string.dashboard_stats).toUpperCase(l);
 			}
 			return null;
-		}
-	}
-
-	public static class BrewStatsFragments extends Fragment {
-
-		private BrewStats brewStats;
-		private List<PlayerStats> playerStats;
-
-		public void setBrewStats(BrewStats brewStats) {
-			this.brewStats = brewStats;
-		}
-
-		public void setPlayerStats(List<PlayerStats> playerStats) {
-			this.playerStats = playerStats;
-		}
-
-		@Override
-		public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-			View rootView = inflater.inflate(R.layout.fragment_dashboard_brew_stats, container, false);
-
-			TextView totalRoundsRun = (TextView) rootView.findViewById(R.id.total_rounds_run);
-			totalRoundsRun.setText("" + brewStats.getTotalTimesRun());
-
-			TextView highestScore = (TextView) rootView.findViewById(R.id.highest_score);
-			highestScore.setText("" + brewStats.getHighestScore());
-
-			TextView lowestScore = (TextView) rootView.findViewById(R.id.lowest_score);
-			lowestScore.setText("" + brewStats.getLowestScore());
-
-			TextView avgPlayerScore = (TextView) rootView.findViewById(R.id.avg_player_score);
-			avgPlayerScore.setText("" + brewStats.getAverageScore());
-
-			TextView avgPlayersPerRound = (TextView) rootView.findViewById(R.id.avg_players_per_round);
-			avgPlayersPerRound.setText("" + brewStats.getAverageNumberOfPlayers());
-
-			return rootView;
-		}
-
-	}
-
-	public static class BrewGroupsFragment extends Fragment implements OnLongClickListener {
-
-		private ExpandableListAdapter adapter;
-		private List<BrewGroup> brewGroups;
-
-		public void setBrewGroups(List<BrewGroup> brewGroups) {
-			this.brewGroups = brewGroups;
-		}
-
-		@Override
-		public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-			View rootView = inflater.inflate(R.layout.fragment_dashboard_brew_groups, container, false);
-			final ExpandableListView listView = (ExpandableListView) rootView.findViewById(R.id.brew_groups_expandable_list_view);
-
-			adapter = new BrewGroupsExpandableListAdapter(brewGroups, getActivity());
-
-			listView.setAdapter(adapter);
-
-			// TODO check this works?
-			// TODO on long click prompt for rename, deletion actions
-			listView.setOnItemLongClickListener(new OnItemLongClickListener() {
-				@Override
-				public boolean onItemLongClick(AdapterView<?> parent, View view, int position, long id) {
-					long packedPosition = listView.getExpandableListPosition(position);
-					if (ExpandableListView.getPackedPositionType(packedPosition) == ExpandableListView.PACKED_POSITION_TYPE_CHILD) {
-						// get item ID's
-						int groupPosition = ExpandableListView.getPackedPositionGroup(packedPosition);
-						int childPosition = ExpandableListView.getPackedPositionChild(packedPosition);
-
-						// handle data
-						LOG.debug("Long Click Group Pos {} Child Pos {}", groupPosition, childPosition);
-
-						// return true as we are handling the event.
-						return true;
-					}
-					return false;
-				}
-			});
-			return rootView;
-		}
-
-		@Override
-		public boolean onLongClick(View view) {
-			return false;
-		}
-	}
-
-	/**
-	 * A dummy fragment representing a section of the app, but that simply
-	 * displays dummy text.
-	 */
-	public static class DummySectionFragment extends Fragment {
-		/**
-		 * The fragment argument representing the section number for this
-		 * fragment.
-		 */
-		public static final String ARG_SECTION_NUMBER = "section_number";
-
-		public DummySectionFragment() {
-		}
-
-		@Override
-		public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-			View rootView = inflater.inflate(R.layout.fragment_dashbaord_dummy, container, false);
-			TextView dummyTextView = (TextView) rootView.findViewById(R.id.section_label);
-			dummyTextView.setText(Integer.toString(getArguments().getInt(ARG_SECTION_NUMBER)));
-			return rootView;
 		}
 	}
